Object Detection/requirements.txt
tensorflow==2.0
numpy
Pillow
opencv-python
seaborn
